ALFMOB-44: Moved some logic from views to view models
