+++ Alfie/Alfie/Views/WishListView/WishListView.swift
@@ -23,21 +23,7 @@ struct WishListView<ViewModel: WishListViewModelProtocol>: View {
             ) {
                 ForEach(viewModel.products) { product in
                     VerticalProductCard(
-                        viewModel: .init(
-                            configuration: .init(size: .medium, hideDetails: false, actionType: .remove),
-                            productId: product.id,
-                            image: product.media.first?.asImage?.url,
-                            designer: product.brand.name,
-                            name: product.name,
-                            priceType: product.priceType,
-                            colorTitle: LocalizableGeneral.$color + ":",
-                            color: product.colour?.name ?? "",
-                            sizeTitle: LocalizableGeneral.$size + ":",
-                            size: product.size == nil ? LocalizableGeneral.$oneSize : product.sizeText,
-                            addToBagTitle: LocalizableGeneral.$addToBag,
-                            outOfStockTitle: LocalizableGeneral.$outOfStock,
-                            isAddToBagDisabled: product.stock == .zero
-                        )
+                        viewModel: viewModel.productCardViewModel(for: product)
                     ) { _, type in
                         handleUserAction(forProduct: product, actionType: type)
                     }
